[LC-458] remove RS
