+++ loopchain/baseservice/object_manager.py
@@ -17,7 +17,6 @@ from loopchain.components import SingletonMetaClass
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from loopchain.radiostation import RadioStationService
     from loopchain.peer import PeerService
     from loopchain.channel.channel_service import ChannelService
 
@@ -26,7 +25,6 @@ class ObjectManager(metaclass=SingletonMetaClass):
     """Provides an interface to reference internal objects.
     """
 
-    rs_service: 'RadioStationService' = None
     peer_service: 'PeerService' = None
     channel_service: 'ChannelService' = None
     rest_proxy_service = None
