+++ loopchain/channel/channel_inner_service.py
@@ -35,7 +35,7 @@ from loopchain.blockchain.votes.v0_1a import BlockVote, LeaderVote
 from loopchain.channel.channel_property import ChannelProperty
 from loopchain.protos import loopchain_pb2, message_code
 from loopchain.qos.qos_controller import QosController, QosCountControl
-from loopchain.rest_server.json_rpc import JsonError
+from loopchain.jsonrpc.exception import JsonError
 from loopchain.utils.message_queue import StubCollection
 
 if TYPE_CHECKING:
