--- loopchain/container/rest_service_rs.py
@@ -1,22 +1,0 @@
-# Copyright 2018 ICON Foundation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""REST Service for Radio station"""
-
-from loopchain.container import ServerType, Container
-
-
-class RestServiceRS(Container):
-    def __init__(self, port):
-        Container.__init__(self, port, ServerType.REST_RS)
-        self.start()  # Container Runs RestServer.start()
