+++ loopchain/peer/__init__.py
@@ -17,6 +17,5 @@ from .status_code import *
 from .block_manager import *
 from .peer_inner_service import *
 from .peer_outer_service import *
-from .channel_manager import *
 from .peer_service import *
 from .consensus_base import *
