+++ loopchain/peer/peer_service.py
@@ -26,16 +26,13 @@ import grpc
 
 from loopchain import configure as conf
 from loopchain import utils
-from loopchain.baseservice import CommonSubprocess, ObjectManager
-from loopchain.container import RestService
+from loopchain.baseservice import CommonSubprocess, ObjectManager, RestService
 from loopchain.crypto.signature import Signer
 from loopchain.peer import PeerInnerService, PeerOuterService
 from loopchain.protos import loopchain_pb2_grpc
 from loopchain.tools.grpc_helper import GRPCHelper
 from loopchain.utils import loggers, command_arguments
 from loopchain.utils.message_queue import StubCollection
-
-DEFAULT_SCORE_PACKAGE = 'score/icx'
 
 
 class PeerService:
