--- loopchain/rest_server/rest_property.py
@@ -1,21 +1,0 @@
-# Copyright 2018 ICON Foundation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from loopchain.components import SingletonMetaClass
-
-
-class RestProperty(metaclass=SingletonMetaClass):
-    def __init__(self):
-        self.node_type = None
-        self.rs_target = None
